+++ pyleco/utils/coordinator_utils.py
@@ -27,20 +27,28 @@ from abc import abstractmethod
 from dataclasses import dataclass
 import logging
 from time import perf_counter
-from typing import Protocol, Optional, Union
+from typing import Any, Protocol, Optional, Union
 
 import zmq
 
 from ..core import COORDINATOR_PORT
-from ..errors import CommunicationError, NOT_SIGNED_IN, DUPLICATE_NAME
 from ..core.message import Message, MessageTypes
 from ..core.serialization import deserialize_data
+from ..json_utils.errors import NOT_SIGNED_IN, DUPLICATE_NAME
 from ..json_utils.rpc_generator import RPCGenerator
 from ..json_utils.json_objects import ErrorResponse, Request
 
 
 log = logging.getLogger(__name__)
 log.addHandler(logging.NullHandler())
+
+
+class CommunicationError(ConnectionError):
+    """Something went wrong, send an `error_msg` to the recipient."""
+
+    def __init__(self, text: str, error_payload: ErrorResponse, *args: Any) -> None:
+        super().__init__(text, *args)
+        self.error_payload = error_payload
 
 
 class MultiSocket(Protocol):