+++ tests/utils/test_coordinator_utils.py
@@ -26,9 +26,10 @@ import pytest
 
 from pyleco.test import FakeContext
 from pyleco.core.message import Message, MessageTypes
-from pyleco.errors import CommunicationError, NOT_SIGNED_IN, DUPLICATE_NAME
-from pyleco.utils.coordinator_utils import ZmqNode, ZmqMultiSocket, Node, Directory, FakeNode
+from pyleco.json_utils.errors import NOT_SIGNED_IN, DUPLICATE_NAME
 from pyleco.json_utils.json_objects import Request, ResultResponse, ErrorResponse
+from pyleco.utils.coordinator_utils import CommunicationError, ZmqNode, ZmqMultiSocket, Node,\
+    Directory, FakeNode
 
 
 class TestZmqMultiSocket: